refactor: move wranglerjs to top level directory
