+++ src/upload/form/mod.rs
@@ -8,11 +8,11 @@ use std::fs;
 use std::path::Path;
 use std::path::PathBuf;
 
-use crate::build::wranglerjs;
 use crate::commands::kv::bucket::AssetManifest;
 use crate::settings::binding;
 use crate::settings::metadata::Metadata;
 use crate::settings::toml::{Target, TargetType};
+use crate::wranglerjs;
 
 use plain_text::PlainText;
 use project_assets::ProjectAssets;
